p4tc: Initialise kernel metadata and make user metadata 32 bit aligned
Add kernel metadata to root pipeline (pipeid 0). Also align user
metadata to 4 bytes.

This commit adds all kernel metadata that is available inside struct
sk_buff. That is:

- pktlen
- datalen
- skbmark
- tcindex
- skbhash
- skbprio
- ifindex
- protocol
- pktype
- idf
- ipsum
- oook
- peeked
- qmap

This commit also adds a fetch operation to metadata, which will be
responsible for fetching the metadata and will be called from the
datapath, more spefically, from metact

The current metadata code has the offset aligned to u8, for example, if
the first metadatum to be added is of size 1, the second metadatum added
would be at offset 1. For performance reasons, we should change the offset
to be 32-bit aligned.

We also are creating the data path structure in the skb, which will be
used to store user metadata and header fields' offsets (more on that in
the very near future).

Fixes: 7936002a1590

Signed-off-by: Victor Nogueira <victor@mojatatu.com>