p4tc: Add init callback for template ops and pipeline init function
Add init callback for template ops so that we can initialise some data
structures during boot time. Specifically, in the case of the pipeline,
we'll need this callback to initialise the root pipeline (pipeid 0).
This special pipeline will hold some kernel specific information, such
as the kernel metadata like skbmark, tcindex and ifindex.

Fixes: 2f8737d25d62

Signed-off-by: Victor Nogueira <victor@mojatatu.com>