p4tc: Add send_port_egress command to metact
Add send_port_egress command to metact, will redirect the packet to egress
of a network interface passed as an operand to the command.

The send_port_egress command has the following syntax:

cmd send_port_egress eth0

Where eth0 is the name of a valid network interface name in the system,
to which we'll redirect the received packet to egress.

================================OPERANDS================================

This commit adds the dev operand

==================================DEV==================================

The DEV operand type represents a network interface inside a Linux system.
We created this new operand so that we could refer to a network interface
device in the send_port_egress command.

To refer to a dev as an operand one must just use its name.

Let's see an example:

sudo tc actions add action metact cmd send_port_egress eth0

Here we have a send_port_egress function that will redirect to device
eth0. This function has a dev operand, which is its only operand, and
in this case is eth0. So, in contrast to the other operands, DEV doesn't
need a path, it's simply the device name (eth0 in our case).

It's important to note that, as of now, a DEV operand can only be used
in command send_port_egress and no place else.

Fixes: 0e29c5e99cb5

Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>