p4tc: Only call parser in the data path when necessary
We should only parse headers in the data path if the pipeline that is
being attached to the P4 filter has header operands being used in its
metact commands.

For example, in the scenario where we just want to filter packets based on
metadata, like skbmark or skbhash, and don't need any header filters,
there's no need to call the parse function in the datapath.

Fixes: 44bb9fd042d2

Co-developed-by: Jamal Hadi Salim <jhs@mojatatu.com>
Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>