p4tc: Remove metact and transform it into dynamic action commands
After some testing and usage of P4TC, we came to the conclusion that
things would be simpler if we removed metact as an action and brought its
commands to dynamic actions. That way, the dynamic actions can directly
use these metact commands instead of having to create a metact, bind
these commands to it and finally point to this new metact directly.

Fixes: 8e297c51ec27

Signed-off-by: Victor Nogueira <victor@mojatatu.com>