p4tc: Add dynamic action's band, bor and bxor commands
================================BAND================================

The band command is used to perform a binary AND operation between two operands
The basic syntax for the band command is:

cmd band operA operB operC

The command will perform the "operB AND operC" and store the result in
operC. That is: operA = operB & operC

operA can only be a metadatum, a key a header field.
operB and operC can only be a constant, a metadatum, a key, a header field
or a param

The following example will perform an AND operation of constant 16 and
mymd metadata and store the result in metadatum mymd2 of pipeline ptables:

tc p4template create action/ptables/myfunc \
   cmd band metadata.ptables.mymd2 metadata.ptables.mymd constant.bit32.16

================================BOR================================

The bor command is used to perform an binary OR operation between two operands
The basic syntax for the bor command is:

cmd bor operA operB operC

The command will perform the "operB OR operC" and store the result in
operC. That is: operA = operB | operC

operA can only be a metadatum, a key a header field.
operB and operC can only be a constant, a metadatum, a key, a header field
or a param

The following example will perform an OR operation of constant 16 and
mymd metadata and store the result in metadatum mymd2 of pipeline ptables:

tc p4template create action/ptables/myfunc \
   cmd bor metadata.ptables.mymd2 metadata.ptables.mymd constant.bit32.16

================================BXOR================================

The bxor command is used to perform an binary XOR operation between two operands
The basic syntax for the bxor command is:

cmd bxor operA operB operC

The command will perform the "operB XOR operC" and store the result in
operC. That is: operA = operB ^ operC

operA can only be a metadatum, a key a header field.
operB and operC can only be a constant, a metadatum, a key, a header field
or a param

The following example will perform a XOR operation of constant 16 and
mymd metadata and store the result in metadatum mymd2 of pipeline ptables:

tc p4template create action/ptables/myfunc \
   cmd bxor metadata.ptables.mymd2 metadata.ptables.mymd constant.bit32.16

Fixes: 8e297c51ec27

Signed-off-by: Evangelos Haleplidis <ehalep@mojatatu.com>
Co-developed-by: Jamal Hadi Salim <jhs@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>