p4tc: Don't reject tclass create command that doesn't specify mandatory arguments
Instead of rejecting a table class create command that doesn't specify
table class postactions or key actions, let's accept and verify when
sealing the pipeline if all the necessary table class arguments are
specified. If key actions or table class post actions or not specified,
we'll not allow the pipeline to be sealed. This will increase usability
by making the table class create command simpler.

Fixes: 2df57c838bc9

Signed-off-by: Victor Nogueira <victor@mojatatu.com>