p4tc: Implement default hit and miss action logic in table apply
Implement default hit action and default miss action logic in table
apply command. That is, the default hit action will be executed whenever a
table lookup results in a match and the table entry which matched has no
action associated with it. The default miss action will be executed
whenever a table lookup doesn't match any of the entries.

Fixes: 8e297c51ec27

Signed-off-by: Victor Nogueira <victor@mojatatu.com>