p4tc: Add 6 more optional operands to concat command
Allow user to concat 8 operands in total, instead of only two
This is intended to improve usability. Before, if the user wished to
concat, for example, 3 operands, the user would have to concat 2, store
the result in an intermediary operand and then concat the last operand
with the result of the previous concat. Now, the user just needs to call
concat passing the 3 operands (4 counting the result operand).
Like the following:

cmd concat opA opB opC opD

With this we also restructured struct p4tc_cmd_operands by removing opA,
opB and opC and making those a part of the operands list

There's also a bug fix in concat and in the assignment of bitstart when
creating a metadata operand.

Fixes: 8e297c51ec27

Signed-off-by: Victor Nogueira <victor@mojatatu.com>