p4tc: Add key as new P4 type
Keys may sometimes be bigger than the maximum allowed type, in our case
u128, that means that we need to create a new type that fits and operates
with values with more that 128 bits. To solve this, we are creating a
key as a new type. Differently from other P4 types, it doesn't have a
fixed size and its size may vary from 0 to P4TC_MAX_KEYSZ (as of now 512
bits). To deal with the fact that the key type may have variable size,
we needed to add the struct p4tc_type into a few of the type's
callbacks, like host_read, host_write and validate_p4t.

Fixes: b89cc771bb6a

Signed-off-by: Victor Nogueira <victor@mojatatu.com>