p4tc: Allow default actions to be updated at runtime
Open an exception in P4TC template and allow specific table attributes
(default hit and miss action) to be updated after pipeline is sealed.
Also add the concept of permissions to default hit and miss actions
The permissions field is a 16bit value which will hold CRUDX (create,
read, update, delete and execute) permissions for control and data path.
Bits 9-5 will have the CRUDX values for control and bits 4-0 will have
CRUDX values for data path. The user can set these permissions when
creating the default actions or when updating them.

For example:

$TC p4template update table/ptables/mytable \
default_miss_action permissions 0x30F action ok

                    CRUDXCRUDX
0x30F in binary is (1100001111), which means we have only Create and
Read permissions in the control plane and Read, Update, Delete and
eXecute permissions in the data plane. This means, for example, that now
we can only delete the default miss action from the data plane. The only
exception is the delete pipeline command, which always will delete the
table and its default actions independent of their permissions.

This PR also introduces a specific delete command for default miss and
hit actions. The command is the following for deleting miss action:

$TC p4template del table/ptables/mytable default_miss_action

The one for deleting hit action is analogous:

$TC p4template del table/ptables/mytable default_hit_action

Fixes: 2df57c838bc9

Signed-off-by: Victor Nogueira <victor@mojatatu.com>