p4tc: Add permissions to tables
Add CRUDX permissions to tables. Mainly what we'll use is the create
permissions. The create permissions for tables will tell us if control
or data plane are allowed to create table entries in a specific table.

For example, the following command will create a table without create
permissions for data and control path:

$TC p4template create table/ptables/cb/tname tblid 1 keysz 64
permissions 0x3C9 ...

Recall that these permissions come in the form of CRUDXCRUDX, where the
first CRUDX block is for control and the last is for data path.

So 0x3C9 is equivalent to CRUD--R--X

If we were to do a get with the following command:

$TC p4template get table/ptables/cb/tname

The output would be the following:

pipeline name ptables
pipeline id 22
    table id 1
    table name cb/tname
    key_sz 64
    max entries 256
    masks 8
    default key 1
    table entries 0
    permissions CRUD--R--X

Fixes: 2df57c838bc9

Signed-off-by: Victor Nogueira <victor@mojatatu.com>