p4tc: Adapt dynamic action commands' to search for names and IDs
Adapt dynamic actions commands' to search for names instead of IDs to
find tables, table keys, pipelines, metadata, parsers, header fields and
registers. This will make the API more versatile and allow us to remove
the commands dependency on the introspection file, since we required
introspection to translate the names the user typed in the command to
IDs. Also add pipeline to struct p4tc_action to help us obtain the
pipeline from the commands without having to search for it yet again.

Fixes: 8e297c51ec27

Signed-off-by: Victor Nogueira <victor@mojatatu.com>