p4tc: Add read-only field to metadata and change metadata.kernel.iif's type to DEV
Change metadata.kernel.iif's type to DEV, so that we can use it to
redirect packets from whence they came. Specifically, it's interesting
to emulate PNA's ingress_spec and egress_spec intrinsic metadata. We
also add a field to the metadata structure to tell if it is a read-only
metadata.

Fixes: 235490dc90d7

Signed-off-by: Victor Nogueira <victor@mojatatu.com>