p4tc: Add support for P4 header validity field
For example, in a P4 program that wants to check if an ethernet header
field was parsed. The user would do something like this:

if (hdr.ethernet.isValid()) {
    ....
}

To emulate this behaviour in P4TC we do the following:

When creating the template header fields object, the user must also
create one corresponding to the validity field. So, for example:

$TC p4template create hdrfield/calc/calc_parser/ethernet/dstAddr \
hdrfieldid 1 parserid 1 type macaddr

$TC p4template create hdrfield/calc/calc_parser/ethernet/srcAddr \
hdrfieldid 2 parserid 1 type macaddr

$TC p4template create hdrfield/calc/calc_parser/ethernet/etherType \
hdrfieldid 3 parserid 1 type bit16

$TC p4template create hdrfield/calc/calc_parser/ethernet/isValid \
hdrfieldid 4 parserid 1 type bit1

In this example, we created the ethernet header fields and additionaly
created the isValid header field. Note that the isValid header field
must be of type bit1.

Now to check if the header ethernet was parsed, the user should do the
following:

...
cmd beq hdrfield.ptables.1.ethernet.isValid constant.bit1.1 control pipe / ok \
...

If header ethernet was parsed, isValid will be 1, otherwise it will be 0.

Fixes: 6328a049dddd

Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Jamal Hadi Salim <jhs@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>