net: sched: act_api: Add dynamic actions IDR
Kernel native actions kinds, such as gact, etc, have a predefined action
ID that is statically defined; example TCA_ID_GACT (XXX) for gact. In P4
we would require to generate new actions "on the fly" based on the P4
program requirement. Doing this in a list, like act_base does, would require us to
us to reimplement all of the ID generation that we get for free with IDRs,
so we chose to stick with an IDR for this. We could've simply converted
act_base into an IDR and used the same structure to store kernel native and
dynamic actions. However dynamic action kinds, like the pipeline they are
attached to, must be per net namespace, as opposed to native action kinds
which are global. For that reason, we chose to create a separate structure
to store dynamic actions.

Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>