net: sched: act_api: Add dynamic actions IDR
Kernel native actions kinds, such as gact, mirred and ife, have a predefine
predefined ID that is given to them statically by the code, like
TCA_ID_GACT. Dynamic actions will require ideas to be generated
"on the fly".Doing this in a list, like act_base does, would require us to
reimplement all of the ID generation that we get for free with IDRs, so we
chose to stick with an IDR for this. We could've simply converted act_base
into an IDR and used the same structure to store kernel native and dynamic
actions. However dynamic action kinds, like the pipeline they are attached
to, must be per net namespace, as opposed to native action kinds which
are global. For that reason, we chose to create a separate structure to
store dynamic actions.

Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>