p4tc: Create a params array to be used in the data path
We already have a params_idr to store the action templates's parameters.
The issue is that using an IDR in the data path results in a performance
penatly. To solve this issue, create a params array, which will be
accessed by the data path operations to retrive a parameter. The
parameters IDR will still be used, but by the the control plane.

Fixes: 068f4a4695e5

Signed-off-by: Victor Nogueira <victor@mojatatu.com>