p4tc: Add tbl_masks_array and tbl_free_masks_bitmap to struct p4tc_table
To lookup for the key in the data path, the algorithmic TCAM iterates
through all the masks in the table and ANDs does masks with the key.
The code stores these masks in an IDR and the looping through the masks
involved calling idr_get_next_entry_ul, which is a costly function.
So to improve data path performance, we are also going to store these
masks in an array, which will make the mask iteration process cheaper.

Fixes: 3cf5983b0386

Signed-off-by: Victor Nogueira <victor@mojatatu.com>