p4tc: Allow optional label on every operation
Allow for an optional label on every operation, instead of introducing a
label operation. For instance, the following P4 action:

action actname() {
    if (metadata.mymd == 1) {
       metadata.mymd2 = 1
    } else {
       metadata.mymd2 = 2
    }
}

Would compile to the following P4TC action:

$TC p4template create action/pname/actname  \
        cmd beq metadata.pname.mymd constant.bit1.1 control pipe / jump else \
        cmd set metadata.mymd2 constant.bit32.1 \
        cmd jump endif \
        cmd label else \
        cmd set metadata.mymd2 constant.bit32.2 \
        cmd label endif

Notice that we needed to create a cmd which simply says: "cmd label else".
Which seems artificial and unnecessary. To solve that, we are enabling
users to specify labels in every command. So the above action would be
simplified into this:

$TC p4template create action/pname/actname  \
        cmd beq metadata.pname.mymd constant.bit1.1 control pipe / jump else \
        cmd set metadata.mymd2 constant.bit32.1 \
        cmd jump endif \
        cmd label else set metadata.mymd2 constant.bit32.2 \
        cmd label endif

Also fix jump calculation bug in branching instructions.

Fixes: a61cfe3cf175

Signed-off-by: Victor Nogueira <victor@mojatatu.com>