p4tc: add P4 classifier
Introduce P4 tc classifier. A  tc filter instantiated on this classifier
is used to bind a P4 pipeline to one or more netdev ports. To use P4
classifier you must specify a pipeline name that will be associated to
this filter. That pipeline must have already been created via a template.
For example, if we were to add a filter to ingress of network interface
device eth0 and associate it to P4 pipeline simple_l3 we'd issue the
following command:

tc filter add dev lo parent ffff: protocol any prio 6 p4 pname simple_l3

The filter itself has the following steps:

================================1 PARSING================================

The parser is implemented in ebpf residing either at the TC or XDP
level.
Note: the different headers and their IDs are already described in the
templating definition and the generated ebpf code.

The parser could be a separate progman or integrated into the ebpf program.
For example, when we have a separate parser program which is loaded into
XDP, we first need to load it into XDP using with the ip command:

ip link set $P0 xdp obj $PROGNAME.o section parser/xdp verbose

Then we pin it:

bpftool prog pin id $ID pin /tmp/

After that we create the P4 filter and reference the XDP program:

$TC filter add dev $P0 ingress protocol ip prio 1 p4 pname redirect_srcip \
    prog xdp pinned /tmp/xdp_parser prog_cookie 22

Note that we also specify a the "prog_cookie", which is used to verify
whether the eBPF program has executed or not before we reach the P4
classifier. The eBPF program sets this cookie by using the kfunc
bpf_p4tc_set_cookie.

=============================FILTER ACTIONS=============================

Filter actions will be where the eBPF program that implements the
remaining P4 program components, aside from the parser, will reside in
the form of a TC BPF action. Of course, the user can choose to not
specify a separate eBPF program as a parser and simply implement all the
P4 program components in one eBPF program, which will be the filter
action.

The filter will return the op code returned by this action.

Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>