p4tc: Add P4 externs
P4 externs are an abstraction in the language to call for external
functionality that is outside of the P4 program itself. For example,
the function that sends a packet to a specific port (send_to_port) in P4
is an extern.

Externs can be seen as classes, which have constructors and methods.
Take, for example, the Register extern definition:

extern Register<T> {
    @tc_construct Register(@tc_numel bit<32> size);
    T read(@tc_key bit<32> index);
    void write(@tc_key bit<32> index, @tc_data T value);
}
XXX: Need to specify read or write notation for each method

A Register is an abstraction for an array in P4, which stores state...

XXX: Here we'll exemplify how externs work in P4

=========================EXTERN TEMPLATE COMMANDS=========================

Now that we've seen how externs work, let's see how we'd translate that
to P4TC commands.

First, we need to create an extern template, which corresponds to the
extern definition in P4. The command would be the following for
register:

tc p4template create extern/root/register ext_id 1 numinstances 2.

The extern template is always associated with the root pipeline, which
is global. So every extern is visible by all pipelines, however each
pipeline will have it's own extern instances. We'll better describe this
once we introduce the template instance command. numinstances specifies
how many instances of this extern will be allowed throughout the user
defined pipelines. ext_id is the id the compiler will give to this
extern.

Note that the above command will also try to load a module called
ext_register.ko. This module will register the extern's ops for the
register extern. It will define a struct p4tc_extern_ops with the
necessary information. We'll discuss this in more detail throughout this
text.

Once we create the extern template, we can define instances of it and
associate them we our user defined pipeline (which correspond to P4
programs). The extern instance is what will define the type of the
extern parameters. For example, let's suppose we wanted to emulate an
instantation of the extern in the P4 program as follows:

Register<bit<32>>(128) reg1;

In this case we have are defining the generic types as bit32 and saying
we'll have 128 elements in this register. The corresponding template
instance creation command will be:

tc p4template create extern_inst/aP4proggie/register/reg1 instid 1 \
numelems 128 \
method read method_id 1 param index type bit32 \
method write method_id 2 param index type bit32 param value type bit32 \
control_path tc_key index type bit32 param value type bit32

We here describe the type of each of the method's parameters and the type
of the control path parameters. Here, we say that the control path index of
this extern will be of type bit32 and the value it will store also be a
bit32.

Note that the instance is always associated with a user defined P4
pipeline and has a name (reg1) and an id (1).

=========================EXTERN RUNTIME COMMANDS=========================

After that, once we seal the pipeline, we can populate those indexes
through runtime commands. For example, if we were to populate index 2
with the value 22, we'd issue the following command:

$TC p4runtime create aP4proggie/extern/register/reg1 tc_key index 2 \
param value 22

We can also update this index with another value:

$TC p4runtime create aP4proggie/extern/register/reg1 tc_key index 2 \
param value 33

Or get its value:

$TC p4runtime get aP4proggie/extern/register/reg1 tc_key index 2

Which will yield the following output:

total exts 0

        extern order 1:
          index  id 1 type bit32 value: 1
          value  id 2 type bit32 value: 33

We can also dump all of the elements in this register:

$TC p4runtime get aP4proggie/extern/register/reg1

============================DATAPATH INTERFACE============================

To call an extern from eBPF, one will the kfunc bpf_skb_p4tc_run_extern.

The kfunc has the following signature:

int
bpf_skb_p4tc_run_extern(struct __sk_buff *skb_ctx,
                        struct p4tc_ext_bpf_params *params,
                        struct p4tc_ext_bpf_res *res)

The first parameter is the skb, the second consists of the extern data
path parameters, which is the following:

struct p4tc_ext_bpf_params {
        u32 pipe_id;
        u32 ext_id;
        u32 inst_id;
        u32 index;
        u32 method_id;
        u32 flags;
        u8 in_params[128]; //extern specific params if any
};

- pipeid: pipeline ID of the extern instance we wish to call
- ext_id: extern ID we wish to call
- inst_id: extern instance ID we wish to call
- index: index of the element we want to access from the control plane
- method_id: method ID we wish to call
- flags: P4TC_EXT_FLAGS_CONTROL_READ or P4TC_EXT_FLAGS_CONTROL_WRITE
- params: The extern method parameters

The third parameter consists of the returned result from the extern,
which has the following definition:

struct p4tc_ext_bpf_res {
        u32 ext_id;
        u32 index_id;
        u32 verdict;
        u32 PAD0;
        u8 params[128];
}

- ext_id: extern ID we called
- inst_id: extern instance ID we called
- verdict: extern verdict, i.e TC_ACT_OK, TC_ACT_PIPE, ...
- params: Extern method return values

This kfunc will end up calling the exec callback defined in the extern
module, unless one of the flags is set. When they are set, they'll
implement either a read from the control plane parameter with the
specified index or a write to the control plane parameter with the
specified index. In case of a write, the implementation will write the
to the index the value specified in &res->params[0].

XXX: Here we should put this kfunc call from a eBPF program.

Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>